Исправил имена классов реализаций интерфейсов, поведение метода SortImpl(Object)f и обработку null
