+++ src/ru/avalon/java/dev/j10/commons/PersonImpl.java
@@ -4,15 +4,15 @@
  * and open the template in the editor.
  */
 package ru.avalon.java.dev.j10.commons;
-import java.util.Date;
+import java.time.*;
 import ru.avalon.java.dev.j10.labs.Person;
 
-public class ClassPerson implements Person {
+public class PersonImpl implements Person {
     
     private String name;
-    private Date birthDate;
+    private LocalDate birthDate;
 
-    public ClassPerson(String name, Date birthDate) {
+    public PersonImpl(String name, LocalDate birthDate) {
         this.name = name;
         this.birthDate = birthDate;
     }
@@ -22,20 +22,17 @@ public class ClassPerson implements Person {
         return name;
     }
 
-    @Override
-    public Date getBirthDate() {
+    @Override 
+    public LocalDate getBirthDate() {
         return birthDate;
     }
     
     @Override
     public int compareTo(Person o) {
-        try{
-            int result = this.name.compareTo(o.getName());
-            if (result == 0) {
-                result = this.birthDate.compareTo(o.getBirthDate());
-            }
-            return result;
+        int result = this.name.compareTo(o.getName());
+        if (result == 0) {
+            result = this.birthDate.compareTo(o.getBirthDate());
         }
-        catch (NullPointerException e){return 0;}
+        return result;   
     }
 }
